+++ hydra-node/src/Hydra/Chain.hs
@@ -126,8 +126,9 @@ data OnChainTx tx
   | OnCollectComTx {headId :: HeadId}
   | OnDepositTx
       { headId :: HeadId
-      , deposited :: UTxOType tx
       , depositTxId :: TxIdType tx
+      , deposited :: UTxOType tx
+      , created :: UTCTime
       , deadline :: UTCTime
       }
   | OnRecoverTx
@@ -202,7 +203,7 @@ data PostTxError tx
   | FailedToConstructCloseTx
   | FailedToConstructContestTx
   | FailedToConstructCollectTx
-  | FailedToConstructDepositTx
+  | FailedToConstructDepositTx {failureReason :: Text}
   | FailedToConstructRecoverTx {failureReason :: Text}
   | FailedToConstructIncrementTx {failureReason :: Text}
   | FailedToConstructDecrementTx {failureReason :: Text}